refactor(base-grid): Barrel file renaming
